✨ add multi-proc(sharing via files) registry
