+++ lib/local_problem_registry/Cargo.toml
@@ -7,4 +7,5 @@ edition = "2021"
 axum = { workspace = true }
 tokio = { workspace = true }
 futures = { workspace = true }
+uuid = { workspace = true }
 judge_core = { path = "../judge_core" }
