+++ lib/local_problem_registry/src/multi_proc.rs
@@ -1,0 +1,2 @@
+pub mod registry_client;
+pub mod registry_server;
